refactor: 스케줄 타입 변경사항 반영
