+++ src/types/schedules.ts
@@ -1,0 +1,5 @@
+export type AvailableSchedule = {
+  date: string;
+  startTime: string;
+  endTime: string;
+};
